move helpers to Runner
