+++ website/src/plugins/rspressPluginPlayground/shared/stringConstants.ts
@@ -1,0 +1,1 @@
+export const APP_FILE_NAME = 'TS'
