+++ website/src/plugins/rspressPluginPlayground/web/Playground/Runner/constants.ts
@@ -1,2 +1,1 @@
-export const APP_FILE_NAME = 'TS'
 export const GET_IMPORT_FUNC_NAME = '__get_import'
